pyproject.toml
[tool.poetry]
name = "not-llama-fs"
version = "0.1.0"
description = ""
authors = ["drforse <george.lifeslice@gmail.com>"]
readme = "README.md"
packages = [{include = "not_llama_fs"}]

[tool.poetry.dependencies]
python = "^3.11"
ollama = "^0.2.1"
asciitree = "^0.3.3"
groq = "^0.9.0"
openai = "^1.34.0"
anthropic = "^0.28.1"
python-magic = "^0.4.27"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
